+++ packages/site/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.4](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.10.3...@jobber/atlantis-site@0.10.4) (2024-12-02)
+
+**Note:** Version bump only for package @jobber/atlantis-site
+
+
+
+
+
+## [0.10.3](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.10.2...@jobber/atlantis-site@0.10.3) (2024-11-29)
+
+**Note:** Version bump only for package @jobber/atlantis-site
+
+
+
+
+
 ## [0.10.2](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.10.1...@jobber/atlantis-site@0.10.2) (2024-11-29)
 
 **Note:** Version bump only for package @jobber/atlantis-site